+++ PCKonfigurator/Gehäuselüfter.cs
@@ -15,7 +15,7 @@ namespace PCKonfigurator
         {            
         }
 
-        public Gehäuselüfter(Int32 _id, String _hersteller, String _typ, Int16 _preis, Double _breite, String _farbe)
+        public Gehäuselüfter(Int32 _id, String _hersteller, String _typ, decimal _preis, Double _breite, String _farbe)
         {
                         this.ID = _id;
             this.Hersteller = _hersteller;
